+++ src/main.ts
@@ -5,6 +5,8 @@ import { Adapter } from "./adapter";
 import { OUTPUT_CHANNEL } from "./core/shared/logger";
 import { UIExtensionMethods } from "./infrastructure/ui-extension-methods";
 
+let testExplorerAdapter: Adapter;
+
 export async function activate(context: vscode.ExtensionContext) {
   const workspaceFolder = (vscode.workspace.workspaceFolders || [])[0];
   const channel = vscode.window.createOutputChannel(OUTPUT_CHANNEL);
@@ -27,7 +29,7 @@ export async function activate(context: vscode.ExtensionContext) {
       context.subscriptions.push(vscode.commands.registerCommand(command, callback));
     };
 
-    const testExplorerAdapter = new Adapter(workspaceFolder, log, channel);
+    testExplorerAdapter = new Adapter(workspaceFolder, log, channel);
     const uiExtensionMethods = new UIExtensionMethods(testExplorerAdapter);
 
     registerCommand("angular-karma-test-explorer.select-project", async () => uiExtensionMethods.createSelectProjectQuickPick());
@@ -35,4 +37,8 @@ export async function activate(context: vscode.ExtensionContext) {
     // this will register an AngularKarmaTestAdapter for each WorkspaceFolder
     context.subscriptions.push(new TestAdapterRegistrar(testHub, workspaceFolder => testExplorerAdapter, log));
   }
+}
+
+export async function deactivate() {
+  testExplorerAdapter.dispose();
 }